MixedLayerThermoclineDynamics

MixedLayerThermoclineDynamics.AbstractField — Type

Abstract supertype for fields.

MixedLayerThermoclineDynamics.AbstractGrid — Type

Abstract supertype for grids.

MixedLayerThermoclineDynamics.AbstractLocation — Type

Abstract supertype for location of variables.

MixedLayerThermoclineDynamics.AbstractTopology — Type

Abstract supertype for topology of grids.

MixedLayerThermoclineDynamics.Bounded — Type

Type for location at the cell centres.

MixedLayerThermoclineDynamics.Centre — Type

Type for location at the cell centres.

MixedLayerThermoclineDynamics.Face — Type

Type for location at the cell faces.

MixedLayerThermoclineDynamics.Field1D — Type

struct Field1D{LX<:AbstractLocation}

A field datatype for 1D objects.

data::AbstractArray

Array with the values of the field.

grid::Any

The grid on which the field lives.

MixedLayerThermoclineDynamics.Field2D — Type

struct Field2D{LX<:AbstractLocation, LY<:AbstractLocation}

A field datatype for 2D objects.

data::Array

Array with the values of the field.

grid::Grid2D

The grid on which the field lives.

MixedLayerThermoclineDynamics.Grid1D — Type

struct Grid1D{Tx<:AbstractTopology} <: AbstractGrid

Returns a one-dimensional staggered grid with topology Tx.

nx::Int64

Number of points in x-direction

hx::Int64

Number of halo points in x-direction

dx::Float64

Grid spacing in x-direction

Lx::Float64

Domain extent in x-direction

xF::AbstractArray

Cell faces in x-direction

xC::AbstractArray

Cell centres in x-direction

MixedLayerThermoclineDynamics.Grid1D — Method

Grid1D(Tx, nx, x_start, x_end; hx=0)

Construct a one-dimensional staggered grid on domain x ∈ [x_start, x_end] with topology Tx, with nx interior grid points, and hx halo points.

MixedLayerThermoclineDynamics.Grid2D — Type

struct Grid2D{Tx<:AbstractTopology, Ty<:AbstractTopology} <: AbstractGrid

Returns a two-dimensional staggered grid with topologies {Tx, Ty}.

nx::Int64

Number of points in x-direction

ny::Int64

Number of points in y-direction

hx::Int64

Number of halo points in x-direction

hy::Int64

Number of halo points in y-direction

dx::Float64

Grid spacing in x-direction

dy::Float64

Grid spacing in y-direction

Lx::Float64

Domain extent in x-direction

Ly::Float64

Domain extent in y-direction

xF::AbstractArray

Cell faces in x-direction

xC::AbstractArray

Cell centres in x-direction

yF::AbstractArray

Cell faces in y-direction

yC::AbstractArray

Cell centres in y-direction

MixedLayerThermoclineDynamics.Grid2D — Method

Grid2D(Tx, Ty, nx, ny, x_start, x_end, y_start, y_end; hx=0, hy=0)

Construct a two-dimensional staggered grid on domain (x, y) ∈ [x_start, x_end] x [y_start, y_end] with topologies {Tx, Ty}, with {nx, ny} interior grid points, and {hx, hy} halo points.

MixedLayerThermoclineDynamics.Periodic — Type

Type for location at the cell centres.

MixedLayerThermoclineDynamics.Field — Method

Field(LX, LY, data, grid::Grid2D)

Constructs a 2D field of data at location (LX, LY) on grid.

MixedLayerThermoclineDynamics.Field — Method

Field(LX, data, grid::Grid1D)

Constructs a 1D field of data at location LX on grid.

MixedLayerThermoclineDynamics.construct_centres — Method

construct_centres(T::AbstractTopology, n, h, d, L, start_center)

Returns the cell centers locations for a dimension with topology T, number of grid points n and h halo points, grid spacing d, and extent L.

MixedLayerThermoclineDynamics.construct_faces — Method

construct_faces(T::AbstractTopology, n, h, d, L)

Returns the face locations for a dimension with topology T, number of grid points n and h halo points, grid spacing d, and extent L.

MixedLayerThermoclineDynamics.fill_halos! — Method

fill_halos!(field::Field1D{<:Any, Grid1D{Periodic}})

Fill halos for a 1D field that lives on Faces with periodic boundary conditions.